+++ evtstrd/exc.py
@@ -6,3 +6,9 @@ class ServerAlreadyRunningError(Exception):
 class DisconnectedError(Exception):
     def __init__(self) -> None:
         super().__init__("connection lost")
+
+
+class PluginError(Exception):
+    def __init__(self, plugin: str, message: str) -> None:
+        super().__init__(f"error in '{plugin}' plugin: {message}")
+        self.plugin = plugin